trying to get routing working
