+++ package.json
@@ -1,5 +1,6 @@
 {
   "name": "tly-virtualpet",
+  "homepage": "./",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
