+++ src/store/selectors/routes.js
@@ -1,8 +1,14 @@
 
 import { getSearchObj } from '@util/tools';
 
-export const selectPetFromUrl = queryString => {
-  console.log('selectPetFromUrl')
+export const selectPetFromWindow = () => {
+  console.log('selectPetFromWindow')
+  const search = window.location.search;
+  return selectPetFromSearchQuery(search);
+}
+
+export const selectPetFromSearchQuery = queryString => {
+  console.log('selectPetFromSearchQuery', queryString)
   if(!queryString) return null;
 
   const obj = getSearchObj(queryString);
